unified-doc-intelligence/requirements.txt
# Core FastAPI and server
fastapi>=0.104.0,<0.115.0
uvicorn[standard]>=0.24.0,<0.35.0
python-multipart>=0.0.6,<0.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0,<2.0.0

# PDF Processing
PyMuPDF>=1.23.0,<1.27.0
pdfplumber>=0.10.0,<0.11.0

# Machine Learning & AI
torch>=2.0.0,<3.0.0
transformers>=4.30.0,<5.0.0
sentence-transformers>=2.3.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
numpy>=1.24.0,<2.0.0
scipy>=1.11.0,<2.0.0

# Vector Search
faiss-cpu>=1.7.0,<2.0.0

# Natural Language Processing
nltk>=3.8.0,<4.0.0

# LangChain
langchain>=0.1.0,<1.0.0
langchain-community>=0.0.20,<1.0.0
langchain-core>=0.1.0,<1.0.0
langchain-text-splitters>=0.0.1,<1.0.0

# Google AI
google-generativeai>=0.7.0,<1.0.0

# Text-to-Speech
azure-cognitiveservices-speech>=1.30.0,<2.0.0
google-cloud-texttospeech>=2.16.0,<3.0.0
pyttsx3>=2.90

# Audio Processing
pydub>=0.25.1

# Database
pysqlite3-binary>=0.5.0,<1.0.0

# Additional utilities
requests>=2.31.0,<3.0.0
aiofiles>=23.0.0,<25.0.0

# System monitoring (optional)
psutil>=5.9.0,<7.0.0

# Development
pytest>=7.0.0,<9.0.0
pytest-asyncio>=0.21.0,<1.0.0
httpx>=0.24.0,<1.0.0
